# openapi.json
{
...
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/v1/users": {
      "post": {
        "operationId": "createUser",
        "summary": "",
        "description": "Create user\nRole required: USER_MANAGER",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The user has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "tags": ["user"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "listUsers",
        "summary": "",
        "description": "List users\nRole required: USER_MANAGER",
        "parameters": [
          {
            "name": "ns",
            "required": false,
            "in": "query",
            "description": "ns",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "ns_scope",
            "required": false,
            "in": "query",
            "description": "ns scope",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "_sort",
            "required": false,
            "in": "query",
            "description": "排序参数",
            "schema": {
              "enum": ["createAt", "-createAt", "updateAt", "-updateAt"],
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": false,
            "in": "query",
            "description": "按id筛选",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "name_like",
            "required": false,
            "in": "query",
            "description": "名称 模糊查询",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "username",
            "required": false,
            "in": "query",
            "description": "用户名",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "username_like",
            "required": false,
            "in": "query",
            "description": "用户名 模糊查询",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "nickname_like",
            "required": false,
            "in": "query",
            "description": "昵称 模糊查询",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "required": false,
            "in": "query",
            "description": "邮箱",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phone",
            "required": false,
            "in": "query",
            "description": "手机号",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "registerRegion",
            "required": false,
            "in": "query",
            "description": "注册地区，存地区编号",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roles",
            "required": false,
            "in": "query",
            "description": "角色",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "dialingPrefix",
            "required": false,
            "in": "query",
            "description": "区号",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "_limit",
            "required": false,
            "in": "query",
            "description": "分页大小",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "_offset",
            "required": false,
            "in": "query",
            "description": "分页偏移",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A paged array of users.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        },
        "tags": ["user"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/auth/v1/users/{userId}": {
      "get": {
        "operationId": "getUser",
        "summary": "",
        "description": "Find user\nRole required: USER_MANAGER",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "User id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The user with expected id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "tags": ["user"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "updateUser",
        "summary": "",
        "description": "Update user\nRole required: USER_MANAGER",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The user updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "tags": ["user"],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "deleteUser",
        "summary": "",
        "description": "Delete user\nRole required: USER_MANAGER",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No content."
          }
        },
        "tags": ["user"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Auth API Server",
    "description": "Auth API for auth service",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "description": "密码",
            "writeOnly": true
          },
          "avatar": {
            "type": "string",
            "description": "头像"
          },
          "intro": {
            "type": "string",
            "description": "简介"
          },
          "data": {
            "type": "string",
            "description": "额外数据"
          },
          "email": {
            "type": "string",
            "description": "邮箱"
          },
          "language": {
            "type": "string",
            "description": "使用语言"
          },
          "lastSeenAt": {
            "format": "date-time",
            "type": "string",
            "description": "最后活跃时间"
          },
          "lastLoginIp": {
            "type": "string",
            "description": "最后登录 IP"
          },
          "nickname": {
            "type": "string",
            "description": "昵称"
          },
          "ns": {
            "type": "string",
            "description": "所属命名空间"
          },
          "phone": {
            "type": "string",
            "description": "手机号"
          },
          "registerIp": {
            "type": "string",
            "description": "注册 IP"
          },
          "registerRegion": {
            "type": "string",
            "description": "注册地区，存地区编号"
          },
          "roles": {
            "description": "角色",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "super": {
            "type": "boolean",
            "description": "是否超级管理员"
          },
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "dialingPrefix": {
            "type": "string",
            "description": "区号"
          }
        },
        "required": ["ns"]
      },
      "IdentityType": {
        "type": "string",
        "description": "认证类型",
        "enum": ["ID"]
      },
      "Identity": {
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/IdentityType"
          },
          "name": {
            "type": "string",
            "description": "真实姓名"
          },
          "verifyAt": {
            "format": "date-time",
            "type": "string",
            "description": "认证时间"
          },
          "verified": {
            "type": "boolean",
            "description": "是否认证通过"
          }
        },
        "required": ["verified"]
      },
      "User": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "description": "密码",
            "writeOnly": true
          },
          "avatar": {
            "type": "string",
            "description": "头像"
          },
          "intro": {
            "type": "string",
            "description": "简介"
          },
          "data": {
            "type": "string",
            "description": "额外数据"
          },
          "email": {
            "type": "string",
            "description": "邮箱"
          },
          "language": {
            "type": "string",
            "description": "使用语言"
          },
          "lastSeenAt": {
            "format": "date-time",
            "type": "string",
            "description": "最后活跃时间"
          },
          "lastLoginIp": {
            "type": "string",
            "description": "最后登录 IP"
          },
          "nickname": {
            "type": "string",
            "description": "昵称"
          },
          "ns": {
            "type": "string",
            "description": "所属命名空间"
          },
          "phone": {
            "type": "string",
            "description": "手机号"
          },
          "registerIp": {
            "type": "string",
            "description": "注册 IP"
          },
          "registerRegion": {
            "type": "string",
            "description": "注册地区，存地区编号"
          },
          "roles": {
            "description": "角色",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "super": {
            "type": "boolean",
            "description": "是否超级管理员"
          },
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "dialingPrefix": {
            "type": "string",
            "description": "区号"
          },
          "identity": {
            "$ref": "#/components/schemas/Identity"
          },
          "id": {
            "type": "string",
            "description": "Entity id"
          },
          "createAt": {
            "format": "date-time",
            "type": "string",
            "description": "Entity created at when"
          },
          "updateAt": {
            "format": "date-time",
            "type": "string",
            "description": "Entity updated at when"
          },
          "createBy": {
            "type": "string",
            "description": "Entity created by who"
          },
          "updateBy": {
            "type": "string",
            "description": "Entity updated by who"
          }
        },
        "required": ["ns", "id"]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "description": "密码",
            "writeOnly": true
          },
          "avatar": {
            "type": "string",
            "description": "头像"
          },
          "intro": {
            "type": "string",
            "description": "简介"
          },
          "data": {
            "type": "string",
            "description": "额外数据"
          },
          "email": {
            "type": "string",
            "description": "邮箱"
          },
          "language": {
            "type": "string",
            "description": "使用语言"
          },
          "lastSeenAt": {
            "format": "date-time",
            "type": "string",
            "description": "最后活跃时间"
          },
          "lastLoginIp": {
            "type": "string",
            "description": "最后登录 IP"
          },
          "nickname": {
            "type": "string",
            "description": "昵称"
          },
          "ns": {
            "type": "string",
            "description": "所属命名空间"
          },
          "phone": {
            "type": "string",
            "description": "手机号"
          },
          "registerIp": {
            "type": "string",
            "description": "注册 IP"
          },
          "registerRegion": {
            "type": "string",
            "description": "注册地区，存地区编号"
          },
          "roles": {
            "description": "角色",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "super": {
            "type": "boolean",
            "description": "是否超级管理员"
          },
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "dialingPrefix": {
            "type": "string",
            "description": "区号"
          }
        }
      },
...
    }
  }
}